Move signature utils to root level
